Free lunch experiment doc and simulation code
Finshed the introduction md of free lunch experiment. Figure out what does the phase problems really refer to (the phase of the signal is different and the phase of the positive feedback is deffernet), and develop a simultion code for cantilever bathed in thermel noise, verifyed our theory, and trying to find a good way to exptend it to freq shift signal.